Add support for tie handling in leaderboards.
* TieRankingLeaderboard: Members with the same score are given
  the same rank.
* CompetitionRankingLeaderboard: Members with the same score will
  be given the same rank. A gap is left in the ranking numbers
  depending on where the next differing score falls.

Also, PEP8'd the codebase and re-organized the structure of the
leaderboard library so that all the base leaderboard code is no
longer in the __init__.py.
