+++ README.md
@@ -172,8 +172,40 @@ highscore_lb.score_for('david')
 
 ### Ranking a member across multiple leaderboards
 
-```ruby
+```python
 highscore_lb.rank_member_across(['highscores', 'more_highscores'], 'david', 50000, { 'member_name': 'david' })
+```
+
+### Alternate leaderboard types
+
+The leaderboard library offers 3 styles of ranking. This is only an issue for members with the same score in a leaderboard.
+
+Default: The `Leaderboard` class uses the default Redis sorted set ordering, whereby different members having the same score are ordered lexicographically. As per the Redis documentation on Redis sorted sets, "The lexicographic ordering used is binary, it compares strings as array of bytes."
+
+Tie ranking: The `TieRankingLeaderboard` subclass of `Leaderboard` allows you to define a leaderboard where members with the same score are given the same rank. For example, members in a leaderboard with the associated scores would have the ranks of:
+
+```
+| member     | score | rank |
+-----------------------------
+| member_1   | 50    | 1    |
+| member_2   | 50    | 1    |
+| member_3   | 30    | 2    |
+| member_4   | 30    | 2    |
+| member_5   | 10    | 3    |
+```
+
+The `TieRankingLeaderboard` accepts one additional option, `ties_namespace` (default: ties), when initializing a new instance of this class. Please note that in its current implementation, the `TieRankingLeaderboard` class uses an additional sorted set to rank the scores, so please keep this in mind when you are doing any capacity planning for Redis with respect to memory usage.
+
+Competition ranking: The `CompetitionRankingLeaderboard` subclass of `Leaderboard` allows you to define a leaderboard where members with the same score will have the same rank, and then a gap is left in the ranking numbers. For example, members in a leaderboard with the associated scores would have the ranks of:
+
+```
+| member     | score | rank |
+-----------------------------
+| member_1   | 50    | 1    |
+| member_2   | 50    | 1    |
+| member_3   | 30    | 3    |
+| member_4   | 30    | 3    |
+| member_5   | 10    | 5    |
 ```
 
 ## Performance Metrics
