+++ test/leaderboard/leaderboard_test.py
@@ -1,5 +1,5 @@
 from redis import Redis, StrictRedis, ConnectionPool
-from leaderboard import Leaderboard
+from leaderboard.leaderboard import Leaderboard
 import unittest
 import time
 import sure
